make validation a utility
